# walk: Send WorkerResults in batches