@tavianator 2023-11-05
```console
tavianator@tachyon $ hyperfine -w2 fd{,-{master,batch}}" -u '^$' /tmp/empty"
Benchmark 1: fd -u '^$' /tmp/empty
  Time (mean ± σ):     143.1 ms ±   5.3 ms    [User: 9.0 ms, System: 132.6 ms]
  Range (min … max):   134.2 ms … 152.7 ms    21 runs
 
Benchmark 2: fd-master -u '^$' /tmp/empty
  Time (mean ± σ):      63.6 ms ±   8.6 ms    [User: 6.0 ms, System: 58.2 ms]
  Range (min … max):    51.7 ms …  80.0 ms    43 runs
 
Benchmark 3: fd-batch -u '^$' /tmp/empty
  Time (mean ± σ):       6.8 ms ±   2.0 ms    [User: 2.6 ms, System: 7.6 ms]
  Range (min … max):     4.3 ms …  12.4 ms    283 runs
 
  Warning: Command took less than 5 ms to complete. Note that the results might be inaccurate because hyperfine can not calibrate the shell startup time much more precise than this limit. You can try to use the `-N`/`--shell=none` option to disable the shell completely.
 
Summary
  fd-batch -u '^$' /tmp/empty ran
    9.33 ± 3.08 times faster than fd-master -u '^$' /tmp/empty
   20.99 ± 6.36 times faster than fd -u '^$' /tmp/empty
```