@tavianator 2023-11-07
Thanks @tmccombs!  I'll let @sharkdp review it too since he has his own benchmarks and he found a fatal flaw in the last attempt :)