@sharkdp 2023-11-08
So I am comparing master @ d62bbbbcd19ab871c6fc166c71043545e6723127 with this branch @ 815b3b1c094949eda38efad7bd987332d8c66ab3. After struggling a bit with thermal throttling affecting the benchmark results, I now have clean results — and they look great!

* I see the expected massive improvement in startup speed (last benchmark below), which is great
* I see a large (20%) performance gain on searches with lots of results, which is great
* I see a small (< 5%) performance gain on searches with few results, which is also great
* I see a medium (6-8%) performance *loss* on command execution benchmarks, which might be acceptable — but maybe something we could still look into?

## `fd` regression benchmark

### No pattern

| Command | Mean [ms] | Min [ms] | Max [ms] | Relative |
|:---|---:|---:|---:|---:|
| `./fd-master --hidden --no-ignore '' '/some/folder'` | 798.4 ± 33.6 | 746.0 | 842.6 | 1.18 ± 0.05 |
| `./fd-1422 --hidden --no-ignore '' '/some/folder'` | 677.7 ± 8.3 | 667.9 | 698.9 | 1.00 |

### Simple pattern

| Command | Mean [s] | Min [s] | Max [s] | Relative |
|:---|---:|---:|---:|---:|
| `./fd-master '.*[0-9]\.jpg$' '/some/folder'` | 1.492 ± 0.006 | 1.483 | 1.504 | 1.01 ± 0.01 |
| `./fd-1422 '.*[0-9]\.jpg$' '/some/folder'` | 1.479 ± 0.008 | 1.466 | 1.494 | 1.00 |

### Simple pattern (-HI)

| Command | Mean [ms] | Min [ms] | Max [ms] | Relative |
|:---|---:|---:|---:|---:|
| `./fd-master -HI '.*[0-9]\.jpg$' '/some/folder'` | 590.3 ± 2.8 | 585.7 | 594.3 | 1.05 ± 0.01 |
| `./fd-1422 -HI '.*[0-9]\.jpg$' '/some/folder'` | 562.8 ± 2.4 | 557.5 | 566.0 | 1.00 |

### File extension

| Command | Mean [ms] | Min [ms] | Max [ms] | Relative |
|:---|---:|---:|---:|---:|
| `./fd-master -HI --extension jpg '' '/some/folder'` | 644.9 ± 2.7 | 640.1 | 647.6 | 1.04 ± 0.01 |
| `./fd-1422 -HI --extension jpg '' '/some/folder'` | 620.1 ± 2.4 | 616.3 | 622.9 | 1.00 |

### File type

| Command | Mean [ms] | Min [ms] | Max [ms] | Relative |
|:---|---:|---:|---:|---:|
| `./fd-master -HI --type l '' '/some/folder'` | 605.1 ± 8.6 | 599.0 | 628.6 | 1.05 ± 0.02 |
| `./fd-1422 -HI --type l '' '/some/folder'` | 575.2 ± 4.0 | 570.2 | 580.7 | 1.00 |

### Command execution

| Command | Mean [s] | Min [s] | Max [s] | Relative |
|:---|---:|---:|---:|---:|
| `./fd-master 'ab' '/some/folder' --exec echo` | 4.711 ± 0.031 | 4.688 | 4.792 | 1.00 |
| `./fd-1422 'ab' '/some/folder' --exec echo` | 5.109 ± 0.036 | 5.057 | 5.176 | 1.08 ± 0.01 |

### Command execution (large output)

| Command | Mean [s] | Min [s] | Max [s] | Relative |
|:---|---:|---:|---:|---:|
| `./fd-master -tf 'ab' '/some/folder' --exec cat` | 4.719 ± 0.028 | 4.682 | 4.779 | 1.00 |
| `./fd-1422 -tf 'ab' '/some/folder' --exec cat` | 5.015 ± 0.053 | 4.915 | 5.087 | 1.06 ± 0.01 |

## Empty folder benchmark

| Command | Mean [ms] | Min [ms] | Max [ms] | Relative |
|:---|---:|---:|---:|---:|
| `./fd-master -u . /tmp/empty` | 28.2 ± 0.6 | 27.3 | 31.2 | 7.65 ± 0.44 |
| `./fd-1422 -u . /tmp/empty` | 3.7 ± 0.2 | 3.5 | 5.3 | 1.00 |