@tavianator 2023-11-08
Yeah I see the `--exec` regression in my benchmarks too.  I'm guessing it's because we have N receiver threads in that case, not just 1, so there's more contention.  I'm testing a fix.