@tavianator 2023-11-08
Actually the problem is not contention, its that results are not evenly distributed to the receivers because the batch sizes vary wildly.  So one `exec::job()` can get way more results to process than the others.  E.g. I just saw this distribution:

```
-x: 0
-x: 0
-x: 0
-x: 3
-x: 2
-x: 33
-x: 45
-x: 55
-x: 63
-x: 66
-x: 47
-x: 73
-x: 116
-x: 133
-x: 132
-x: 115
-x: 125
-x: 151
-x: 186
-x: 421
```

Maybe lowering the max batch size for `-x` will help.