@tavianator 2023-11-08
> Maybe lowering the max batch size for `-x` will help.

Yep, that works!  `--exec` perf is now better than `master`, and nothing else seems to have regressed.  New benchmark results are in the top comment.