@tmccombs 2023-11-08
I was  actually thinking that for the `--exec` case, it might be more performant to execute on the sender thread instead of having N receiver threads. I'm not sure if that would be worth the complexity though.