@tavianator 2023-11-08
> Unfortunately, it looks like things got worse for me with [1469bf3](https://github.com/sharkdp/fd/commit/1469bf3757b6025224d88817905445c3b124d445)

Interesting!  Doesn't reproduce for me, even with a similar test (`echo` instead of `true`, no `-u`, similar execution time):

```console
tavianator@tachyon $ hyperfine -w2 fd-{batch,master}" --search-path ~/code/linux -d4 -x echo"
Benchmark 1: fd-batch --search-path ~/code/linux -d4 -x echo
  Time (mean ± σ):      3.622 s ±  0.012 s    [User: 36.389 s, System: 32.154 s]
  Range (min … max):    3.602 s …  3.640 s    10 runs
 
Benchmark 2: fd-master --search-path ~/code/linux -d4 -x echo
  Time (mean ± σ):      3.693 s ±  0.011 s    [User: 36.529 s, System: 32.544 s]
  Range (min … max):    3.674 s …  3.707 s    10 runs
 
Summary
  fd-batch --search-path ~/code/linux -d4 -x echo ran
    1.02 ± 0.00 times faster than fd-master --search-path ~/code/linux -d4 -x echo
```

> I was also wondering what happens in case of (few search results but) longer running programs? Would batching lead to situations where one receiver would run significantly longer than the rest?

In the degenerate case, any batch size N > 1 can lead to an Nx slowdown: if there are exactly N matches, all ending up in the same batch (unlikely but possible), they will be executed sequentially rather than in parallel.

Would you mind trying some variations?

- Set the batch size to 1 in `--exec` mode (line 453)
- Set the channel capacity to `2 * config.threads` (line 641)
- Both of the above

Neither of those makes a big difference for me, but they may help on your machine.